Redesign: only export bao runner; re-organize steps and factory function.
